fruit_store/requirements.txt
asgiref==3.3.4
certifi==2020.12.5
Django==2.2
djangorestframework==3.12.4
pytz==2021.1
pymongo
djongo
click
